Implemented BaseCreature within all AI/Items/Traps/Collectibles & MainDude

+++ source/hud/hud.hpp
@@ -30,7 +30,7 @@ public:
 
     bool holding_item_shopping{};
     const char *holding_item_name{};
-    int *holding_item_cost{};
+    u16 *holding_item_cost{};
 
     void disable_all_prompts();
